complete patient pages
